// sdks/node-ts/src/apache_beam/proto/beam_runner_api.ts
export interface PTransform {
    /**
     * (Required) A unique name for the application node.
     *
     * Ideally, this should be stable over multiple evolutions of a pipeline
     * for the purposes of logging and associating pipeline state with a node,
     * etc.
     *
     * If it is not stable, then the runner decides what will happen. But, most
     * importantly, it must always be here and be unique, even if it is
     * autogenerated.
     *
     * @generated from protobuf field: string unique_name = 5;
     */
    uniqueName: string;
    /**
     * (Optional) A URN and payload that, together, fully defined the semantics
     * of this transform.
     *
     * If absent, this must be an "anonymous" composite transform.
     *
     * For primitive transform in the Runner API, this is required, and the
     * payloads are well-defined messages. When the URN indicates ParDo it
     * is a ParDoPayload, and so on. For some special composite transforms,
     * the payload is also officially defined. See StandardPTransforms for
     * details.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.FunctionSpec spec = 1;
     */
    spec?: FunctionSpec;
    /**
     * (Optional) A list of the ids of transforms that it contains.
     *
     * Primitive transforms (see StandardPTransforms.Primitives) are not allowed
     * to specify subtransforms.
     *
     * Note that a composite transform may have zero subtransforms as long as it
     * only outputs PCollections that are in its inputs.
     *
     * @generated from protobuf field: repeated string subtransforms = 2;
     */
    subtransforms: string[];
    /**
     * (Required) A map from local names of inputs (unique only with this map, and
     * likely embedded in the transform payload and serialized user code) to
     * PCollection ids.
     *
     * The payload for this transform may clarify the relationship of these
     * inputs. For example:
     *
     *  - for a Flatten transform they are merged
     *  - for a ParDo transform, some may be side inputs
     *
     * All inputs are recorded here so that the topological ordering of
     * the graph is consistent whether or not the payload is understood.
     *
     * @generated from protobuf field: map<string, string> inputs = 3;
     */
    inputs: {
        [key: string]: string;
    };
    /**
     * (Required) A map from local names of outputs (unique only within this map,
     * and likely embedded in the transform payload and serialized user code)
     * to PCollection ids.
     *
     * The URN or payload for this transform node may clarify the type and
     * relationship of these outputs. For example:
     *
     *  - for a ParDo transform, these are tags on PCollections, which will be
     *    embedded in the DoFn.
     *
     * @generated from protobuf field: map<string, string> outputs = 4;
     */
    outputs: {
        [key: string]: string;
    };
    /**
     * (Optional) Static display data for this PTransform application. If
     * there is none, it may be omitted.
     *
     * @generated from protobuf field: repeated org.apache.beam.model.pipeline.v1.DisplayData display_data = 6;
     */
    displayData: DisplayData[];
    /**
     * Environment where the current PTransform should be executed in.
     *
     * Transforms that are required to be implemented by a runner must omit this.
     * All other transforms are required to specify this.
     *
     * @generated from protobuf field: string environment_id = 7;
     */
    environmentId: string;
    /**
     * (Optional) A map from URNs designating a type of annotation, to the
     * annotation in binary format. For example, an annotation could indicate
     * that this PTransform has specific privacy properties.
     *
     * A runner MAY ignore types of annotations it doesn't understand. Therefore
     * annotations MUST NOT be used for metadata that can affect correct
     * execution of the transform.
     *
     * @generated from protobuf field: map<string, bytes> annotations = 8;
     */
    annotations: {
        [key: string]: Uint8Array;
    };
}
// ...
export interface ParDoPayload {
    /**
     * (Required) The FunctionSpec of the DoFn.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.FunctionSpec do_fn = 1;
     */
    doFn?: FunctionSpec;
    /**
     * (Optional) A mapping of local input names to side inputs, describing
     * the expected access pattern.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.SideInput> side_inputs = 3;
     */
    sideInputs: {
        [key: string]: SideInput;
    };
    /**
     * (Optional) A mapping of local state names to state specifications.
     * If this is set, the stateful processing requirement should also
     * be placed in the pipeline requirements.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.StateSpec> state_specs = 4;
     */
    stateSpecs: {
        [key: string]: StateSpec;
    };
    /**
     * (Optional) A mapping of local timer family names to timer family
     * specifications. If this is set, the stateful processing requirement should
     * also be placed in the pipeline requirements.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.TimerFamilySpec> timer_family_specs = 9;
     */
    timerFamilySpecs: {
        [key: string]: TimerFamilySpec;
    };
    /**
     * (Optional) Only set when this ParDo contains a splittable DoFn.
     * If this is set, the corresponding standard requirement should also
     * be placed in the pipeline requirements.
     *
     * @generated from protobuf field: string restriction_coder_id = 7;
     */
    restrictionCoderId: string;
    /**
     * (Optional) Only set when this ParDo can request bundle finalization.
     * If this is set, the corresponding standard requirement should also
     * be placed in the pipeline requirements.
     *
     * @generated from protobuf field: bool requests_finalization = 8;
     */
    requestsFinalization: boolean;
    /**
     * Whether this stage requires time sorted input.
     * If this is set, the corresponding standard requirement should also
     * be placed in the pipeline requirements.
     *
     * @generated from protobuf field: bool requires_time_sorted_input = 10;
     */
    requiresTimeSortedInput: boolean;
    /**
     * Whether this stage requires stable input.
     * If this is set, the corresponding standard requirement should also
     * be placed in the pipeline requirements.
     *
     * @generated from protobuf field: bool requires_stable_input = 11;
     */
    requiresStableInput: boolean;
}
/**
 * @generated from protobuf message org.apache.beam.model.pipeline.v1.StateSpec
 */
export interface StateSpec {
    /**
     * @generated from protobuf oneof: spec
     */
    spec: {
        oneofKind: "readModifyWriteSpec";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ReadModifyWriteStateSpec read_modify_write_spec = 1;
         */
        readModifyWriteSpec: ReadModifyWriteStateSpec;
    } | {
        oneofKind: "bagSpec";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.BagStateSpec bag_spec = 2;
         */
        bagSpec: BagStateSpec;
    } | {
        oneofKind: "combiningSpec";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.CombiningStateSpec combining_spec = 3;
         */
        combiningSpec: CombiningStateSpec;
    } | {
        oneofKind: "mapSpec";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.MapStateSpec map_spec = 4;
         */
        mapSpec: MapStateSpec;
    } | {
        oneofKind: "setSpec";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.SetStateSpec set_spec = 5;
         */
        setSpec: SetStateSpec;
    } | {
        oneofKind: "orderedListSpec";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.OrderedListStateSpec ordered_list_spec = 6;
         */
        orderedListSpec: OrderedListStateSpec;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message org.apache.beam.model.pipeline.v1.ReadModifyWriteStateSpec
 */
export interface ReadModifyWriteStateSpec {
    /**
     * @generated from protobuf field: string coder_id = 1;
     */
    coderId: string;
}
/**
 * @generated from protobuf message org.apache.beam.model.pipeline.v1.BagStateSpec
 */
export interface BagStateSpec {
    /**
     * @generated from protobuf field: string element_coder_id = 1;
     */
    elementCoderId: string;
}
// ...
export interface TimerFamilySpec {
    /**
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.TimeDomain.Enum time_domain = 1;
     */
    timeDomain: TimeDomain_Enum;
    /**
     * @generated from protobuf field: string timer_family_coder_id = 2;
     */
    timerFamilyCoderId: string;
}
// ...
export interface SideInput {
    /**
     * (Required) URN of the access pattern required by the `view_fn` to present
     * the desired SDK-specific interface to a UDF.
     *
     * This access pattern defines the SDK harness <-> Runner Harness RPC
     * interface for accessing a side input.
     *
     * The only access pattern intended for Beam, because of its superior
     * performance possibilities, is "beam:sideinput:multimap" (or some such
     * URN)
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.FunctionSpec access_pattern = 1;
     */
    accessPattern?: FunctionSpec;
    /**
     * (Required) The FunctionSpec of the UDF that adapts a particular
     * access_pattern to a user-facing view type.
     *
     * For example, View.asSingleton() may include a `view_fn` that adapts a
     * specially-designed multimap to a single value per window.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.FunctionSpec view_fn = 2;
     */
    viewFn?: FunctionSpec;
    /**
     * (Required) The FunctionSpec of the UDF that maps a main input window
     * to a side input window.
     *
     * For example, when the main input is in fixed windows of one hour, this
     * can specify that the side input should be accessed according to the day
     * in which that hour falls.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.FunctionSpec window_mapping_fn = 3;
     */
    windowMappingFn?: FunctionSpec;
}
// ...
export interface FunctionSpec {
    /**
     * (Required) A URN that describes the accompanying payload.
     * For any URN that is not recognized (by whomever is inspecting
     * it) the parameter payload should be treated as opaque and
     * passed as-is.
     *
     * @generated from protobuf field: string urn = 1;
     */
    urn: string;
    /**
     * (Optional) The data specifying any parameters to the URN. If
     * the URN does not require any arguments, this may be omitted.
     *
     * @generated from protobuf field: bytes payload = 3;
     */
    payload: Uint8Array;
}
// ...
export interface DisplayData {
    /**
     * A key used to describe the type of display data. See StandardDisplayData
     * for the set of well known urns describing how the payload is meant to be
     * interpreted.
     *
     * @generated from protobuf field: string urn = 1;
     */
    urn: string;
    /**
     * (Optional) The data specifying any parameters to the URN. If
     * the URN does not require any arguments, this may be omitted.
     *
     * @generated from protobuf field: bytes payload = 2;
     */
    payload: Uint8Array;
}
// ...
export const PTransform = new PTransform$Type();
// ...
export const ParDoPayload = new ParDoPayload$Type();
// ...
export const StateSpec = new StateSpec$Type();
// ...
export const ReadModifyWriteStateSpec = new ReadModifyWriteStateSpec$Type();
// ...
export const BagStateSpec = new BagStateSpec$Type();
// ...
export const MapStateSpec = new MapStateSpec$Type();
// ...
export const TimerFamilySpec = new TimerFamilySpec$Type();
// ...
export const SideInput = new SideInput$Type();
// ...
export const FunctionSpec = new FunctionSpec$Type();
// ...
export const DisplayData = new DisplayData$Type();