// sdks/node-ts/src/apache_beam/proto/beam_fn_api.ts
export interface RemoteGrpcPort {
    /**
     * (Required) An API descriptor which describes where to
     * connect to including any authentication that is required.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ApiServiceDescriptor api_service_descriptor = 1;
     */
    apiServiceDescriptor?: ApiServiceDescriptor;
    /**
     * (Required) The ID of the Coder that will be used to encode and decode data
     * sent over this port.
     *
     * @generated from protobuf field: string coder_id = 2;
     */
    coderId: string;
}
// ...
export interface ProcessBundleDescriptor {
    /**
     * (Required) A pipeline level unique id which can be used as a reference to
     * refer to this.
     *
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * (Required) A map from pipeline-scoped id to PTransform.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.PTransform> transforms = 2;
     */
    transforms: {
        [key: string]: PTransform;
    };
    /**
     * (Required) A map from pipeline-scoped id to PCollection.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.PCollection> pcollections = 3;
     */
    pcollections: {
        [key: string]: PCollection;
    };
    /**
     * (Required) A map from pipeline-scoped id to WindowingStrategy.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.WindowingStrategy> windowing_strategies = 4;
     */
    windowingStrategies: {
        [key: string]: WindowingStrategy;
    };
    /**
     * (Required) A map from pipeline-scoped id to Coder.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.Coder> coders = 5;
     */
    coders: {
        [key: string]: Coder;
    };
    /**
     * (Required) A map from pipeline-scoped id to Environment.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.Environment> environments = 6;
     */
    environments: {
        [key: string]: Environment;
    };
    /**
     * A descriptor describing the end point to use for State API
     * calls. Required if the Runner intends to send remote references over the
     * data plane or if any of the transforms rely on user state or side inputs.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ApiServiceDescriptor state_api_service_descriptor = 7;
     */
    stateApiServiceDescriptor?: ApiServiceDescriptor;
    /**
     * A descriptor describing the end point to use for Data API for user timers.
     * Required if the ProcessBundleDescriptor contains any transforms that have user timers.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ApiServiceDescriptor timer_api_service_descriptor = 8;
     */
    timerApiServiceDescriptor?: ApiServiceDescriptor;
}
// ...
class RemoteGrpcPort$Type extends MessageType<RemoteGrpcPort> {
    constructor() {
        super("org.apache.beam.model.fn_execution.v1.RemoteGrpcPort", [
            { no: 1, name: "api_service_descriptor", kind: "message", T: () => ApiServiceDescriptor },
            { no: 2, name: "coder_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<RemoteGrpcPort>): RemoteGrpcPort {
        const message = { coderId: "" };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<RemoteGrpcPort>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: RemoteGrpcPort): RemoteGrpcPort {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* org.apache.beam.model.pipeline.v1.ApiServiceDescriptor api_service_descriptor */ 1:
                    message.apiServiceDescriptor = ApiServiceDescriptor.internalBinaryRead(reader, reader.uint32(), options, message.apiServiceDescriptor);
                    break;
                case /* string coder_id */ 2:
                    message.coderId = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: RemoteGrpcPort, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* org.apache.beam.model.pipeline.v1.ApiServiceDescriptor api_service_descriptor = 1; */
        if (message.apiServiceDescriptor)
            ApiServiceDescriptor.internalBinaryWrite(message.apiServiceDescriptor, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* string coder_id = 2; */
        if (message.coderId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.coderId);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message org.apache.beam.model.fn_execution.v1.RemoteGrpcPort
 */
export const RemoteGrpcPort = new RemoteGrpcPort$Type();
// ...
export const ProcessBundleDescriptor = new ProcessBundleDescriptor$Type();